requirements.txt
torch
torchvision
streamlit
numpy
matplotlib
scikit-learn
pillow
